backup/requirements.txt
functions-framework==3.*
google-auth==2.29.0
requests==2.31.0
